requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
hypercorn==0.18.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.44
alembic==1.12.1

# File Upload
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database Drivers (choose one based on your database)
# PostgreSQL
# psycopg2-binary==2.9.9

# MySQL
# pymysql==1.1.0

# SQLite (included in Python standard library)

# Email validation
email-validator==2.3.0

# Additional utilities
python-dateutil==2.8.2
